+++ client/src/components/FormControl/FormControl.module.scss
@@ -1,4 +1,8 @@
 .formControl {
+  --text-color-dark: var(--primary-text);
+  --text-color-light: var(--secondary-text);
+  --border-color: var(--separator);
+
   position: relative;
   width: 100%;
