+++ client/src/index.scss
@@ -35,6 +35,8 @@ html {
 
   --separator: #dbdbdb;
   --success: #58c322;
+  --link: #00376B;
+  --blue-facebook: #385185;
 }
 
 // Mixins
