+++ client/src/pages/Login/Login.module.scss
@@ -2,51 +2,43 @@
   --background-color: var(--primary-background);
   --text-color-dark: var(--primary-text);
   --text-color-light: var(--secondary-text);
-  --link-color: var(--primary-button);
   --border-color: var(--separator);
+  --link-color: var(--link);
+  --primary-blue: var(--primary-button);
+  --blue: var(--blue-facebook);
 
-  margin-top: 10px;
   text-align: center;
   font-size: 14px;
   max-width: 350px;
+  width: 350px;
   display: flex;
   flex-direction: column;
   align-items: center;
   gap: 10px;
 
-  
   .formContainer {
+    width: 100%;
     border: 1px solid var(--border-color);
     display: flex;
     flex-direction: column;
     align-items: center;
     gap: 10px;
-    
+
     .logo {
-      margin: 50px 0 10px 0;
+      margin: 50px 0 20px 0;
     }
 
     .form {
+      width: 100%;
       display: flex;
       flex-direction: column;
       align-items: center;
-      padding: 0 40px 40px 40px;
+      padding: 0 40px 25px 40px;
       gap: 6px;
 
-      .mainText {
-        color: var(--text-color-light);
-        font-weight: 700;
-        font-size: 16px;
-      }
-
-      .facebookBtn {
+      .loginBtn {
         width: 100%;
-        margin: 10px 0;
-        height: 34px;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        gap: 5px;
+        margin-top: 20px;
       }
 
       .or {
@@ -78,14 +70,27 @@
         }
       }
 
-      .signupBtn {
-        width: 100%;
-        margin-top: 20px;
+      .facebookBtn {
+        color: var(--blue);
+        text-decoration: none;
+        font-weight: 700;
+        height: 34px;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        gap: 6px;
+      }
+
+      .forgotPwdLink {
+        margin-top: 8px;
+        color: var(--link-color);
+        text-decoration: none;
+        font-size: 12px;
       }
     }
   }
 
-  .loginPrompt {
+  .signupPrompt {
     border: 1px solid var(--border-color);
     width: 100%;
     height: 70px;
@@ -93,9 +98,10 @@
     align-items: center;
     justify-content: center;
 
-    .login {
+    .signup {
       text-decoration: none;
-      color: var(--link-color);
+      font-weight: 700;
+      color: var(--primary-blue);
     }
   }
 }
