+++ client/src/routes/app.router.js
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Home from '../pages/Home';
-import Login from '../pages/Login';
+import Login from '../pages/Login/Login';
 import Register from '../pages/Register/Register';
 
 function AppRouter() {
